Delete the qwen2 tokenizer, use the default huggingface tokenizer
