+++ tests/test_qwen2_tokenizer.py
@@ -6,20 +6,20 @@ from mlx_lm.tokenizer_utils import TokenizerWrapper
 
 # 更新导入路径指向 src 目录下的模块
 from mlx_omni_server.schemas.chat_schema import Role
-from mlx_omni_server.services.chat.tools.qwen2 import Qwen2ChatTokenizer
+from mlx_omni_server.services.chat.tools.hugging_face import HuggingFaceChatTokenizer
 
 
 class TestQwen2ChatTokenizer(unittest.TestCase):
     def setUp(self):
         mock_tokenizer = Mock(spec=TokenizerWrapper)
-        self.qwen2_tokenizer = Qwen2ChatTokenizer(mock_tokenizer)
+        self.hf_tokenizer = HuggingFaceChatTokenizer(mock_tokenizer)
 
     def test_decode_single_tool_call(self):
         # Test single tool call with double quotes
         text = """<tool_call>
 {"name": "get_current_weather", "arguments": {"location": "Boston, MA", "unit": "fahrenheit"}}
 </tool_call>"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertEqual(result.role, Role.ASSISTANT)
@@ -39,7 +39,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
         text = """<tool_call>
 {"name": "get_current_weather", invalid_json}
 </tool_call>"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertEqual(result.role, Role.ASSISTANT)
@@ -53,7 +53,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
         text = """<tool_call>
 {"arguments": {"location": "Boston, MA"}}
 </tool_call>"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertEqual(result.role, Role.ASSISTANT)
@@ -65,7 +65,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
     def test_decode_non_tool_call(self):
         # Test non-tool call text
         text = "This is a regular message without any tool calls."
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertEqual(result.role, Role.ASSISTANT)
@@ -74,7 +74,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
 
     def test_loose_mode_with_response_tag(self):
         # 启用宽松模式
-        self.qwen2_tokenizer.strict_mode = False
+        self.hf_tokenizer.strict_mode = False
 
         # Test with <response> tag
         text = """<response>
@@ -88,7 +88,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
         #   {"name": "get_current_weather", "arguments": {"location": "Boston, MA", "unit": "celsius"}}
         # </function-calls>
         #         """
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
         print(f"=======\nresult: \n{result}")
 
         self.assertIsNotNone(result)
@@ -104,13 +104,13 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
 
     def test_loose_mode_with_function_tag(self):
         # 启用宽松模式
-        self.qwen2_tokenizer.strict_mode = False
+        self.hf_tokenizer.strict_mode = False
 
         # Test with <function-calls> tag
         text = """<function-calls>
           {"name": "get_current_weather", "arguments": {"location": "Boston, MA", "unit": "celsius"}}
         </function-calls>"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
         print(f"=======\nresult: \n{result}")
 
         self.assertIsNotNone(result)
@@ -126,13 +126,13 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
 
     def test_loose_mode_with_markdown(self):
         # 启用宽松模式
-        self.qwen2_tokenizer.strict_mode = False
+        self.hf_tokenizer.strict_mode = False
 
         # Test with markdown code block
         text = """```xml
         {"name": "get_current_weather", "arguments": {"location": "Boston, MA", "unit": "fahrenheit"}}
         ```"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertIsNotNone(result.tool_calls)
@@ -141,7 +141,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
 
     def test_loose_mode_with_xml_declaration(self):
         # 启用宽松模式
-        self.qwen2_tokenizer.strict_mode = False
+        self.hf_tokenizer.strict_mode = False
 
         # Test with XML declaration
         text = """<?xml version="1.0" encoding="UTF-8"?>
@@ -151,7 +151,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
             "arguments": {"location": "Boston, MA", "unit": "celsius"}
         }
         </json>"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertIsNotNone(result.tool_calls)
@@ -164,7 +164,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
 
     def test_loose_mode_with_reversed_fields(self):
         # 启用宽松模式
-        self.qwen2_tokenizer.strict_mode = False
+        self.hf_tokenizer.strict_mode = False
 
         # Test with reversed fields order (arguments before name)
         text = """<tools>
@@ -173,7 +173,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
             "name": "get_current_weather"
         }
         </tools>"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertIsNotNone(result.tool_calls)
@@ -186,7 +186,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
 
     def test_strict_mode_rejects_loose_format(self):
         # 确保严格模式下拒绝非标准格式
-        self.qwen2_tokenizer.strict_mode = True
+        self.hf_tokenizer.strict_mode = True
 
         # Test with <response> tag (should fail in strict mode)
         text = """<response>
@@ -195,7 +195,7 @@ class TestQwen2ChatTokenizer(unittest.TestCase):
           "arguments": {"location": "Boston, MA", "unit": "fahrenheit"}
         }
         </response>"""
-        result = self.qwen2_tokenizer.decode(text)
+        result = self.hf_tokenizer.decode(text)
 
         self.assertIsNotNone(result)
         self.assertEqual(result.role, Role.ASSISTANT)
